+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-26T12:03:42.938Z")
+    created: Optional[datetime] = parse_datetime("2025-09-29T03:53:49.945Z")
     r"""The date and time the resource was created"""