+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-26T12:03:39.765Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-29T03:53:46.885Z")
     r"""The date and time the resource was last updated"""